backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
email-validator>=2.0.0
motor>=3.3.0
pymongo>=4.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1
python-multipart>=0.0.9
openai>=1.30.0
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
pinecone-client>=3.0.0
httpx>=0.27.0
parsedatetime>=2.6
